+++ bindgen/gen_rust.py
@@ -5,39 +5,16 @@
 #   - types are PascalCase
 #   - otherwise snake_case
 # -------------------------------------------------------------------------------
-import gen_ir
-import os, shutil, sys
+import os, sys
 
 import gen_util as util
 
-module_names = {
-    "slog_": "log",
-    "sg_": "gfx",
-    "sapp_": "app",
-    "stm_": "time",
-    "saudio_": "audio",
-    "sgl_": "gl",
-    "sdtx_": "debugtext",
-    "sshape_": "shape",
-    "simgui_": "imgui",
-    "sglue_": "glue",
-}
+lib_root = 'sokol-rust/src'
+module_root = f'{lib_root}/sokol'
+c_root = f'{module_root}/c'
 
 module_requires_rust_feature = {
-    module_names["simgui_"]: "imgui",
-}
-
-c_source_paths = {
-    "slog_": "sokol-rust/src/sokol/c/sokol_log.c",
-    "sg_": "sokol-rust/src/sokol/c/sokol_gfx.c",
-    "sapp_": "sokol-rust/src/sokol/c/sokol_app.c",
-    "stm_": "sokol-rust/src/sokol/c/sokol_time.c",
-    "saudio_": "sokol-rust/src/sokol/c/sokol_audio.c",
-    "sgl_": "sokol-rust/src/sokol/c/sokol_gl.c",
-    "sdtx_": "sokol-rust/src/sokol/c/sokol_debugtext.c",
-    "sshape_": "sokol-rust/src/sokol/c/sokol_shape.c",
-    "simgui_": "sokol-rust/src/sokol/c/sokol_imgui.c",
-    "sglue_": "sokol-rust/src/sokol/c/sokol_glue.c",
+    'imgui': 'imgui',
 }
 
 ignores = [
@@ -739,7 +716,9 @@ def pre_parse(inp):
             for item in decl["items"]:
                 enum_items[enum_name].append(as_enum_item_name(item["name"]))
 
-def gen_imports(inp, dep_prefixes):
+def gen_imports(inp):
+    dep_prefixes = inp['dep_prefixes']
+    module_names = inp['module_names']
     for dep_prefix in dep_prefixes:
         dep_module_name = module_names[dep_prefix]
         # l(f'const {dep_prefix[:-1]} = @import("{dep_module_name}.rs");')
@@ -806,7 +785,7 @@ def gen_helpers(inp):
     #     l("")
 
 
-def gen_module(inp, dep_prefixes):
+def gen_module(inp):
     module = inp['module']
     if module in module_requires_rust_feature:
         feature = module_requires_rust_feature[module]
@@ -819,7 +798,7 @@ def gen_module(inp, dep_prefixes):
     l("#![allow(dead_code)]")
     l("#![allow(unused_imports)]")
     l("")
-    gen_imports(inp, dep_prefixes)
+    gen_imports(inp)
     gen_helpers(inp)
     pre_parse(inp)
     prefix = inp["prefix"]
@@ -852,37 +831,24 @@ def gen_module(inp, dep_prefixes):
 
 
 def prepare():
-    print("=== Generating Rust bindings:")
-    if not os.path.isdir("sokol-rust/src/sokol"):
-        os.makedirs("sokol-rust/src/sokol")
-    if not os.path.isdir("sokol-rust/src/sokol/c"):
-        os.makedirs("sokol-rust/src/sokol/c")
-
-    with open("sokol-rust/src/lib.rs", "w", newline="\n") as f_outp:
-        f_outp.write("//! Automatically generated sokol bindings for Rust\n\n")
+    util.prepare('Rust', module_root, c_root)
+    with open(f'{lib_root}/lib.rs', 'w', newline='\n') as f_outp:
+        f_outp.write('//! Automatically generated sokol bindings for Rust\n\n')
 
 
-def gen(c_header_path, c_prefix, dep_c_prefixes):
-    if c_prefix not in module_names:
-        print(f' >> warning: skipping generation for {c_prefix} prefix...')
-        return
-
-    module_name = module_names[c_prefix]
-    c_source_path = c_source_paths[c_prefix]
-    print(f'  {c_header_path} => {module_name}')
+def gen(opts):
     reset_globals()
-    c_path_in_project = f'sokol-rust/src/sokol/c/{os.path.basename(c_header_path)}'
-    shutil.copyfile(c_header_path, c_path_in_project)
-    ir = gen_ir.gen(c_header_path, c_source_path, module_name, c_prefix, dep_c_prefixes)
-    gen_module(ir, dep_c_prefixes)
-    output_path = f"sokol-rust/src/{ir['module']}.rs"
-    with open(output_path, 'w', newline='\n') as f_outp:
-        f_outp.write(out_lines)
+    success, ir = util.gen_ir(opts, c_root)
+    if success:
+        gen_module(ir)
+        output_path = f"sokol-rust/src/{ir['module']}.rs"
+        with open(output_path, 'w', newline='\n') as f_outp:
+            f_outp.write(out_lines)
 
-    with open("sokol-rust/src/lib.rs", "a", newline="\n") as f_outp:
-        module = ir['module']
-        if module in module_requires_rust_feature:
-            feature = module_requires_rust_feature[module]
-            f_outp.write(f"/// Enable feature \"{feature}\" to use\n")
-            f_outp.write(f"#[cfg(feature=\"{feature}\")]\n")
-        f_outp.write(f"pub mod {module};\n")
+        with open("sokol-rust/src/lib.rs", "a", newline="\n") as f_outp:
+            module = ir['module']
+            if module in module_requires_rust_feature:
+                feature = module_requires_rust_feature[module]
+                f_outp.write(f"/// Enable feature \"{feature}\" to use\n")
+                f_outp.write(f"#[cfg(feature=\"{feature}\")]\n")
+            f_outp.write(f"pub mod {module};\n")