+++ bindgen/gen_util.py
@@ -1,5 +1,6 @@
 # common utility functions for all bindings generators
-import re
+import re, shutil, os
+import gen_ir as ir
 
 re_1d_array = re.compile(r"^(?:const )?\w*\s*\*?\[\d*\]$")
 re_2d_array = re.compile(r"^(?:const )?\w*\s*\*?\[\d*\]\[\d*\]$")
@@ -55,3 +56,28 @@ def as_lower_camel_case(s, prefix):
     for part in parts[1:]:
         outp += part.capitalize()
     return outp
+
+def prepare(language_name, rel_mod_root_dir, rel_c_root_dir):
+    print(f'=== Generating {language_name} bindings:')
+    if not os.path.isdir(rel_mod_root_dir):
+        os.makedirs(rel_mod_root_dir, exist_ok = True)
+    if not os.path.isdir(rel_c_root_dir):
+        os.makedirs(rel_c_root_dir, exist_ok = True)
+    shutil.copyfile('impl/sokol_defines.h', f'{rel_c_root_dir}/sokol_defines.h')
+
+def gen_ir(opts, c_root, with_comments=False):
+    c_src_header_path = opts['c_header_path']
+    c_prefix = opts['c_prefix']
+    dep_c_prefixes = opts['dep_c_prefixes']
+    module_names = opts['module_names']
+    if not c_prefix in module_names:
+        print('>> warning: skipping generation for {c_prefix} prefix...')
+        return False, {}
+    c_src_source_path = f'impl/{os.path.splitext(os.path.basename(c_src_header_path))[0]}.c'
+    c_dst_header_path = f'{c_root}/{os.path.basename(c_src_header_path)}'
+    c_dst_source_path = f'{c_root}/{os.path.splitext(os.path.basename(c_src_header_path))[0]}.c'
+    shutil.copyfile(c_src_header_path, c_dst_header_path)
+    shutil.copyfile(c_src_source_path, c_dst_source_path)
+    module_name = module_names[c_prefix]
+    res = ir.gen(c_dst_header_path, c_dst_source_path, module_name, module_names, c_prefix, dep_c_prefixes, with_comments)
+    return True, res