+++ bindgen/impl/sokol_app.c
@@ -1,0 +1,5 @@
+#if defined(IMPL)
+#define SOKOL_APP_IMPL
+#endif
+#include "sokol_defines.h"
+#include "sokol_app.h"