+++ bindgen/impl/sokol_args.c
@@ -1,0 +1,5 @@
+#if defined(IMPL)
+#define SOKOL_ARGS_IMPL
+#endif
+#include "sokol_defines.h"
+#include "sokol_args.h"