+++ bindgen/impl/sokol_fetch.c
@@ -1,0 +1,5 @@
+#if defined(IMPL)
+#define SOKOL_FETCH_IMPL
+#endif
+#include "sokol_defines.h"
+#include "sokol_fetch.h"