+++ bindgen/impl/sokol_gfx.c
@@ -1,0 +1,5 @@
+#if defined(IMPL)
+#define SOKOL_GFX_IMPL
+#endif
+#include "sokol_defines.h"
+#include "sokol_gfx.h"